Add ping latency metric
Adds timing for the ping redis command, which can provide a more
accurate measure of the actual latency with less extra overhead.

This also changes the timer for the info latency from `process_time`
to a proper monotonic wall clock, because the time we want to measure
is mainly I/O which is "sleep time" which wouldn't be reflected on
`process_time`.