First iteration of diagnose API for integrations
This introduces the first design idea, providing an internal object
for integrations to easily access diagnose-related functionality, such
as reporting diagnoses as well as defining their own explicit
diagnostics. This removes the need to subclass `get_diagnoses()` (the
agent-facing API) and provides a convenient namespace for all the
functionality, without requiring explicit imports.